Write PCA calculations to output transformed data
Add remaining backend functionality (minus the actual API endpoints)
to perform dimensionality reduction on the audio features of songs
in a Spotify playlist such that each song is plottable in 3D space.
Song metadata and coordinates (x, y, z) are stored in a pandas
DataFrame but will be sent to the frontend in JSON format.